+++ packages/lib/passesBoundaryCheck/index.ts
@@ -2,11 +2,18 @@ import { Map } from 'ol'
 import { Coordinate } from 'ol/coordinate'
 import VectorLayer from 'ol/layer/Vector'
 import VectorSource from 'ol/source/Vector'
+import { LayerBoundPluginOptions } from '@polar/lib-custom-types'
 
 // arbitrarily give up after 10s of stalling
 let readinessChecks = 0
 const readinessCheckLimit = 100
 const readinessWaitTime = 100
+
+export const errors = {
+  undefinedBoundaryLayer: Symbol.for('Boundary Layer undefined'),
+  undefinedBoundarySource: Symbol.for('Boundary Source undefined'),
+  sourceNotReady: Symbol.for('Source not ready'),
+} as const
 
 /**
  * @param source - source to check
@@ -33,9 +40,9 @@ const isReady = async (source: VectorSource): Promise<boolean> => {
  */
 export const passesBoundaryCheck = async (
   map: Map,
-  boundaryLayerId: string | undefined,
+  boundaryLayerId: LayerBoundPluginOptions['boundaryLayerId'],
   coordinate: Coordinate
-): Promise<boolean> => {
+): Promise<boolean | symbol> => {
   if (typeof boundaryLayerId === 'undefined') {
     return Promise.resolve(true)
   }
@@ -51,7 +58,7 @@ export const passesBoundaryCheck = async (
     console.error(
       `POLAR Map Client: No layer configured to match boundaryLayerId "${boundaryLayerId}" in plugins/GeoLocation/src/utils/isWithinBoundary.ts.`
     )
-    return Promise.resolve(false)
+    return Promise.resolve(errors.undefinedBoundaryLayer)
   }
 
   const boundaryLayerSource = boundaryLayer.getSource()
@@ -60,7 +67,7 @@ export const passesBoundaryCheck = async (
     console.error(
       `POLAR Map Client: Layer with boundaryLayerId "${boundaryLayerId}" missing source in plugins/GeoLocation/src/utils/isWithinBoundary.ts.`
     )
-    return Promise.resolve(false)
+    return Promise.resolve(errors.undefinedBoundarySource)
   }
 
   const sourceReady = await isReady(boundaryLayerSource)
@@ -69,7 +76,7 @@ export const passesBoundaryCheck = async (
     console.error(
       `POLAR Map Client: Layer with boundaryLayerId "${boundaryLayerId}" did not load or is featureless in plugins/GeoLocation/src/utils/isWithinBoundary.ts.`
     )
-    return Promise.resolve(false)
+    return Promise.resolve(errors.sourceNotReady)
   }
 
   const features = boundaryLayerSource.getFeatures() || []