+++ packages/plugins/GeoLocation/src/types.ts
@@ -2,6 +2,7 @@ import Geolocation from 'ol/Geolocation.js'
 import { Vector as VectorSource } from 'ol/source'
 import VectorLayer from 'ol/layer/Vector'
 import Feature from 'ol/Feature'
+import { GeoLocationConfiguration } from '@polar/lib-custom-types'
 export interface GeoLocationState {
   geolocation: Geolocation | null
   position: number[]
@@ -10,12 +11,14 @@ export interface GeoLocationState {
 }
 
 export interface GeoLocationGetters extends GeoLocationState {
-  boundaryLayerId: string | undefined
+  boundaryLayerId: GeoLocationConfiguration['boundaryLayerId']
+  boundaryOnError: GeoLocationConfiguration['boundaryOnError']
   toastAction: string | undefined
   configuredEpsg: string
   checkLocationInitially: boolean
   geoLocationMarkerLayer: VectorLayer<VectorSource>
   keepCentered: boolean
   markerFeature: Feature
+  showTooltip: boolean
   zoomLevel: number
 }