+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.1.0
+
+- Feature: Add core state variable for map's center position.
+
 ## 1.0.0
 
 Initial release.