+++ packages/plugins/GeoLocation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.1.0
+
+- Feature: Update`@polar/lib-passes-boundary-check` to major version 2.
+- Feature: Add `boundaryOnError` parameter to let user define behaviour on boundary check errors.
+- Feature: Add optional tooltip to user geolocation map icon.
+
 ## 1.0.0
 
 Initial release.