+++ packages/plugins/IconMenu/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.0.1
+
+- Fix: Plugin contents were rendered above map in certain circumstances. This has been resolved.
+
 ## 1.0.0
 
 Initial release.