+++ packages/types/custom/CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## 1.1.0
+
+- Feature: Add parameter `boundaryOnError` to `PinsConfiguration`.
+- Feature: Add parameter `boundaryOnError` to `GeoLocationConfiguration`.
+- Feature: Add flag `showTooltip` to `GeoLocationConfiguration`.
+- Feature: Add type for CoreState's `center` field.
+- Feature: Remodel type structure to deduplicate fields now modeled in `LayerBoundPluginOptions`.
+- Feature: Add new optional parameter `initial` to `PinsConfiguration` including related interface `InitialPin`.
+- Feature: Added new configuration variation for `movable` in `PinsConfiguration`.
+- Fix: Add type for CoreState's nested `.plugin` container.
+
 ## 1.0.0
 
 Initial release.